split binary from library
