+++ crates/containerd-shim-wasm/Cargo.toml
@@ -1,9 +1,10 @@
 [package]
-name = "containerd-shim-wasmtime-v1"
+name = "containerd-shim-wasm"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+doctest = false
 
 [dependencies]
 containerd-shim = "0.3.0"
@@ -33,11 +34,3 @@ tempfile = "3.0"
 [features]
 default = []
 generate_bindings = ["ttrpc-codegen"]
-
-[[bin]]
-name = "containerd-shim-wasmtimed-v1"
-path = "src/containerd-shim-wasmtimed-v1/main.rs"
-
-[[bin]]
-name = "containerd-wasmtimed"
-path = "src/containerd-wasmtimed/main.rs"
