+++ crates/containerd-shim-wasm/protos/sandbox.proto
