+++ crates/containerd-shim-wasm/src/sandbox/mod.rs
@@ -6,7 +6,7 @@ pub mod manager;
 pub mod shim;
 
 pub use error::Error;
-pub use instance::{Instance, InstanceConfig, Nop as NopInstance, Wasi as WasiInstance};
+pub use instance::{EngineGetter, Instance, InstanceConfig};
 pub use manager::{Sandbox as SandboxService, Service as ManagerService};
 pub use shim::{Cli as ShimCli, Local};
