--- crates/containerd-shim-wasmtime-v1/src/containerd-shim-wasmtimed-v1/main.rs
@@ -1,6 +1,0 @@
-use containerd_shim as shim;
-use containerd_shim_wasmtime_v1::sandbox::manager::Shim;
-
-fn main() {
-    shim::run::<Shim>("containerd-shim-wasmtimed-v1", None)
-}
